requirements.txt
appdirs==1.4.4
asgiref==3.5.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
decorator==5.1.1
Django==4.0.6
django-filter==22.1
djangorestframework==3.13.1
dogpile.cache==1.1.8
idna==3.3
iso8601==1.0.2
jmespath==1.0.1
jsonpatch==1.32
jsonpointer==2.3
keystoneauth1==5.0.0
Markdown==3.4.1
munch==2.5.0
netifaces==0.11.0
openstacksdk==0.100.0
os-service-types==1.7.0
pbr==5.9.0
pycparser==2.21
pytz==2022.1
PyYAML==6.0
requests==2.28.1
requestsexceptions==1.4.0
six==1.16.0
sqlparse==0.4.2
stevedore==4.0.0
urllib3==1.26.11
